chore: Toolbar 파일에 있던 types를 types.ts로 이동
