+++ client/src/components/_common/Editor/EditorToolbar.tsx
@@ -12,19 +12,12 @@ import {
 import { useEditorContext } from './EditorContext';
 import { useEditorHandlerContext } from './EditorHandlerContext';
 import { TextColorPicker } from './EditorTextColorPicker';
-import type { ToolbarProps } from './types';
+import type { Align, Format, List, Option, Size, TextColor, ToolbarProps } from './types';
 import { applyAlignment } from './utils/alignment';
 import { applyList } from './utils/list';
 import { insertDivider } from './utils/options';
 import { getCurrentFormats, getCurrentLists, getCurrentSize } from './utils/selection';
 import { applyStyle } from './utils/textStyles';
-
-export type Size = '10px' | '12px' | '14px' | '16px' | '24px' | '36px';
-export type Format = 'bold' | 'italic' | 'underline' | 'strikethrough';
-export type Align = 'left' | 'center' | 'right' | 'justify';
-export type List = 'disc' | 'decimal';
-export type Option = 'link' | 'image' | 'divider';
-export type TextColor = 'color' | 'background';
 
 function EditorToolbar({ radius, sx }: ToolbarProps) {
     // prop destruction
